Changed fields from integer to float in the sql file, so we have a floating point number instead of the rounded integer.

+++ htdocs/brasilncm/sql/llx_brasilncm.sql
@@ -23,8 +23,8 @@ create table llx_brasil_ncm
 	  fk_customcode   integer,
 	  ncm_nr          integer NOT NULL,
 	  ncm_descr       varchar(255),
-	  imp_import      integer,
-	  ipi             integer,
-	  pis             integer,
-	  cofins          integer
+	  imp_import      float,
+	  ipi             float,
+	  pis             float,
+	  cofins          float
 )ENGINE=innodb;
